skbuff_debug: don't update free stack until the state is confirmed to be deactive
If the stack is updated, just the second free stack (another victim)
can be seen. With this change the free stack won't get updated until
the state of the SKB has been confirmed.

The free stack for the second free is still available in the backtrace
of the BUG() output.

Change-Id: I94271f3e03786437c07b23f42c0dcb0591a88836
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Signed-off-by: Casey Chen <kexinc@codeaurora.org>