skb_recycler: Add an API to free skb recycler list
Currently we have consume_skb which takes a single skb buffer
as an input and process on them one by one. To avoid this processing
and local_irq calls, we add an API to take a list of skbs as an input
to free them at once to save CPU cycles.
Also, we have avoided global and spare list processing here by assuming
hot skb recycle list is always available.

Change-Id: I5c99b9281ec6d7ab189da4e8d30c6cb8a8817d7c
Signed-off-by: Neha Bisht <quic_nbisht@quicinc.com>