clk: qcom: gcc: ipq5332: add CLK_IGNORE_UNUSED for GCC_MEM_NOC_AHB_CLK
Currently, there is no user for this clock, beacuse of this clock
framework disables this clock. This clock is required for the MEM NOC
XPUs as well. Since the clock is getting disabled, XPUs are also getting
disabled, which is not expected. So lets add the flag CLK_IGNORE_UNUSED
to it.

Change-Id: I1a0092e3a4c5fd16614d1cc6d8e8c9e7c96ed5c4
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>