arm64: dts: qcom: ipq5332: add PCIe PHYs and controller nodes
Add PCIE0, PCIE1, PCIE2 root complex devices along with the corresponding
PHYs found on IPQ5332 SoC. PCIE0 and PCIE2 are single lane RC, PCIE1 is
dual lane RC.

PCIE1 and PCIE2 shared the same PHY. By default, PCIE1 will be operating
in dual lane mode. If we need to use the PCIE2, we can enable the mux,
so that both PCIE1 and PCIE2 will be operating in single lane mode.

Change-Id: I86543ab42d8f90fd7cdeef5d16bb932b818b009a
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>