net: Add DSCP remarking feature for HW acceleration.
This patch includes combination of 5 patches from 5.4 kernel.

1. DSCP remarking is and advanced QoS feature of acceleration
engine. With this feature it is possible to set the accelerated
eggress packets' DSCP values with iptables command.

2. Add Ingress qostag fields in CT extension.

Extend DSCP remark extension structure to accomodate
ingress qostag fields.

3. net: Add DSCP remark set flags

These flags indicates that which QoS features are set
in the DSCP conntrack extension. DSCP, Priority, or IGS.

4. net: add mark to the dscp extension

mark field is needed when we do dscp

5. net: Add DSCP extension MARK flag

Change-Id: If113b2528ce4b9f31615df36bc04d1f72501c321
Signed-off-by: Pavithra R <pavir@codeaurora.org>
Signed-off-by: Ken Zhu <quic_guigenz@quicinc.com>
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>