arm64: dts: qcom: include the GPLL0 as clock provider for mailbox
While the kernel is booting up, APSS PLL will be running at 800MHz with GPLL0
as source. Once the cpufreq driver is available, APSS PLL will be configured
to the rate based on the opp table and the source also will be changed to
APSS_PLL_EARLY. Without this patch, CPU Freq driver reports that CPU is
running at 24MHz instead of the 800MHz.

Also, for IPQ53xx to support the dynamic scaling, GPLL0 is used as the safe
source while switching the frequencies.

Change-Id: If410292251985938e0a322e488a5b643565c9ba5
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>