drivers: remoteproc: q6v5_mpd: Add recovery support for multi-pd
In case of root-pd crash in a multi-pd target, each of the userpds need
to be stopped and started again.

Add support to loop through child nodes and stop each userpd in rootpd
stop function. Add support to spawn a kthread to start all userpds once
rootpd is up in case of rootpd crash.

Change-Id: I49b003bb5ce147ca7a4c3da07518530c807154dc
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>