drivers: mhitest: Add Fix mplat_g[] handling in mhi_test
On PCIe remove, mhi_test doesn't removes that particular instance and
it always deletes the last entry in mplat_g[].
On PCIe rescan, it always tries to insert at the last index in mplat_g[],
which can overwrite the previous entry and leads to crash or QMI failures.
Fixing this by removing that particular instance on PCIs remove and adds that
particular instance on PCIs rescan.

Change-Id: Id84b1e242da618bbcb14b56f69df52f3e9364dcd
Signed-off-by: Poovendhan Selvaraj <quic_poovendh@quicinc.com>