arm64: dts: qcom: ipq9574: Add support for Discrete SMPS on AL06
AL06 does not have PMIC, it has discrete SMPS and no voltage
scaling supported. So this patch disables the voltage scaling
and adds 4state GPIO regulator support.

For APC two GPIOs are used to scale the voltage. apc_rail1 works
based on the part and apc_rail2 works based on the CPU frequency.
CX and MX are combined to one rail which is controlled by cx_rail
with one GPIO.

--------------------------------------------------------------------------
| Regulator |  Fuse  | Threshold |   Mode    | GPIO_1 | GPIO_0 | Voltage |
--------------------------------------------------------------------------
|    APC    |  STUB  |    1V     | S,SL1,N   |   0    |   0    |  0.85V  |
|           |        |           | S,SL1,N   |   0    |   1    |  0.863V |
|           |        |           | NL1,T,TL1 |   1    |   0    |  1.004V |
|           |        |           | NL1,T,TL1 |   1    |   1    |  1.068V |
|           |        |           |           |        |        |         |
|    CX     | MX NOM |   0.85V   |    NA     |   0    |   NA   |  0.85V  |
|           |        |           |    NA     |   1    |   NA   |  0.910V |
--------------------------------------------------------------------------

Also, power rail configuration changes are there in QCN9224 on AL06.
Updates the boot-args for the same.

Change-Id: I53254b08f2b0ec4bdbe49627d11ff351c503fb04
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>