bus: mhi: test: Ignore duplicate RDDM event from MHI
Due to the commit [1], MHI threaded IRQ handler does not filter
notifications for same event to the driver. When the target hits a
syserr, it will send 2 interrupts to the host, one for EE change and one
for state change respectively. This ends up going to the driver as 2
RDDM notifications and driver queues up recovery twice for the same
target leading to undesirable issues.

Ignore duplicate RDDM notification if the current mhi EE and device EE
are the same.

[1]: https://lore.kernel.org/all/1681142292-27571-2-git-send-email-quic_jhugo@quicinc.com/

Change-Id: I072c4fd28f68afdeda217a93e0f0b36c8b41e686
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>