soc: qcom: lm: Support license file from the DTS
Support license file name from DTS to have backward
compatibility. If license-file property with file
name is defined in the DTS node, then it will be used.
conf file will be skipped in that case.

This is a snapshot of the commit
1865114a6647fc209f28b1d3db7c490d82ba1d2c
from win.linuxopenwrt.2.1 branch and
oss/kernel/linux-ipq-5.4 project.

Change-Id: I3bccbe0f1de3fa7e950abbb55979f669fcea1a15
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>