drivers: remoteproc: q6v5_mpd: Protect userpd recovery under lock
When userpd start is in progress, there is a possibility that rootpd can
assert in some cases, for e.g. User-PD auth failure. In this scenario,
recovery is queued for the rootpd from the Fatal/WDOG IRQ handler and
the recovery workq proceeds to do a stop of all userpds and then rootpd.
This causes an secure watchdog crash from TZ as userpd start is still in
progress and rootpd stop is requested.

Protect userpd rproc_stop in q6_wcss_stop and userpd rproc_start in
handle_upd_in_rpd_crash under rproc->lock mutex to avoid such issues on
rootpd recovery when userpd boot is in progress.

Change-Id: Ia07133d0c6fee17fe99ff0db6255544282423f17
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>
Signed-off-by: Sridharan S N <quic_sridsn@quicinc.com>