soc: qcom: lm: Add ioctl support to get Bindings
Add GET_BINDINGS ioctl support to get the bindings
from userspace. Userspace needs to pass the 34 Bytes
of NONCE and 2KB buffer for bindings. LM driver will
get the bindings from TMEL and copy it to the ECDSA
buffer in the response.

Userspace needs to pass the proper sizes in
nonce_buf_len and ecdsa_buf_len.

This is a snapshot of the commit
248cf70145ebe91c44a03fa6bc45f141afb29255
from win.linuxopenwrt.2.1 branch and
oss/kernel/linux-ipq-5.4 project.

Change-Id: Ib0ab8ab4cead1237dab69b6740c13bc5e531559d
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>