drivers: bus: mhi: test: Move mplat list array to linked list
mplat_g global array contains the list of mplats for each probed mhitest
device. This causes issues where one of the device is removed and
re-probed.

Move mplat_g to linked list. This ensures the list entries are not
disturbed on remove/probe of any device and also removes the limit on
the number of devices that can be probed.

While at it, remove and cleanup unused functions which were using the
mplat_g array.

Change-Id: Ifd87107168ba0bcb9a5d3b7fce8901a8eb4b94f2
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>