phy: qcom: pcie-uniphy: Configure multiplexed PCIe PHY for 2lane mode
Avoid considering the multiplexed PCIe PHY will be in 2lane mode by default.
If TCSR handle is present in the DTSI, set it to the default 2lane mode
which helps overwrite the bootloader settings.

This is a snapshot of the commit
3ce1216219ed5b0326f4aeb1145d80ab3f5e3d0e
from win.linuxopenwrt.2.1 branch and
oss/kernel/linux-ipq-5.4 project.

Change-Id: Iba53f8298ac4a7dc7ce7fd123466fbafea1f3a9b
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>