arm64: dts: qcom: devsoc: Add few reserved memory regions
Reserve memory regions for Bootloader, SBL, TZ and TZapp.

Bootloader will collect the system RAM contents upon crash for
the post morterm analysis. If we don't reserve the memory region used by
bootloader, obviously linux will consume it. Upon the next boot on crash,
bootloader will be loaded in the same region, which will lead to loss of data
and sometimes we may even miss out the critical information. So lets
reserve the region used by the bootloader.

Similarly SBL copies some data into the reserved region and it will be
used in the crash scenario. So reserve 1MB for SBL as well.

Change-Id: Ib6b956a23c544f5d1c97a88c9132f170245cd768
Signed-off-by: Devi Priya <quic_devipriy@quicinc.com>