soc: qcom: ipq-debug: add support to capture and log the internal Q6 crash
When the integrated Q6 crashes, we need to update the IMEM location with
value 0xB, so that bootloaders and HLOS can parse and log the reason for
the system crash.

To achieve this, register a atomic SSR notifier, which will be called
whenever the Q6 crashes and update the crash reason in the IMEM
location.

In case of Q6 recovery is enabled, we need to clear the crash reason. To
do so, register a SSR notifier and clear the reason in IMEM when rproc
state is QCOM_SSR_BEFORE_POWERUP. For the cases, where the recovery is
not enabled, client drivers will invoke BUG_ON. In those cases, we need
to retain the crash reason as 0xB.

Q6 will crash either due to FATAL or WDT. In case of FATAL errors, Q6
will write the reason in the IMEM region dedicated to it. In case of
WDT, Q6 will not have control to write into the IMEM region. So while
parsing, if the crash reason is 0xB, check for the IMEM address
dedicated to the Q6, if there is any non-zero value, treat it as FATAL
error, else it is WDT error.

Change-Id: Ic39e6f0e714427857109cda3ef4e176297b189f2
Signed-off-by: Kathiravan Thirumoorthy <quic_kathirav@quicinc.com>
Signed-off-by: Sridharan S N <quic_sridsn@quicinc.com>