platform: ipq: handle the error scenarios in sec_dat
Add fix in sec-upgrade driver to handle all the fuse status properly in sec_dat.

If unwittingly, a non negative return value is returned in the error
paths of store_sec_dat()i and store_sec_auth, the kernel and/or userland
will call store_sec_dat() again assuming that the data was partially written.

Signed-off-by: Poovendhan Selvaraj <quic_poovendh@quicinc.com>
Change-Id: I0b254349b5f007f368bb8e476ef8779ab112d1ad