arm64: DTS: qcom: ipq5332: Add dts property for umac reset interrupt
For QCN6432 the umac reset interupt line is reserved so
adding entry for the same in rdp442, rdp446, rdp479, rdp477,
and rdp481 related dts files.

Also this dedicated interrupt line needs to be resetted everytime
after the interrupt is received so that the next interrupt can be
raised so adding a new entry qcom,interrupt-reset-addr which holds
a slot specific address that should be used to write the reset value
0x00000100 so that the interrupt line will be resetted.

GIC_SPI - 0
Shared peripheral interrupt which means that the interrupt can be
routed to any CPU for servicing.

IRQ_TYPE_EDGE_RISING - 1
Indicates a rising edge triggered Interrupt.

Change-Id: I2221f9872cbbac878213aafb804877dfbb8ef913
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>