drivers: remoteproc: q6v5_upd: Handle rootpd crash during userpd start
If rootpd crashes while userpd spawn is in progress, userpd spawn
interrupt will not be received and till the spawn operation times out
(5s), rootpd will be tried for recovery, and userpd also will be tried
to spawn as previous spawn smp2p bit is not cleared. This will cause
rootpd to crash again in a loop till the first spawn times out.

Register for atomic notifiers for rootpd rproc and complete pending
waits for userpd if NOTIFY_CRASH event is received for the rootpd rproc.

Change-Id: I27ce05a7933901cfad249e9dd2426c17f58cda0e
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>