platform: ipq: Update the sequence to extract metadata size from ELF
The current design is implemented considering that the NULL segments
will be found only at the start of the ELF before the loadable segments.
ATF images are found to have a NULL segment appended at the end of the ELF
which leads to miscalculation of meta data size.
Hence, update the sequence to calculate metadata size accordingly.

While at it, return ld_seg_cnt when we are unable to parse the loadable
segments to avoid infinite looping.

Change-Id: I364629215cdfc0b1f99cf0c28bb4d9c5513db030
Signed-off-by: Devi Priya <quic_devipriy@quicinc.com>