platform: ipq: Update the metadata size calculation sequence when parsing the elf for loadable segments
The current design is implemented considering that the NULL segments
will be found only at the start of the ELF before the loadable segments.
Some images are found to have a NULL segment appended at the end of the ELF
which leads to miscalculation of meta data size.
Hence, update the sequence to calculate metadata size accordingly.

Change-Id: Ic52aeb2a545270816dafc1a9accbe6b7fbf585ca
Signed-off-by: Devi Priya <quic_devipriy@quicinc.com>