arm64: dts: qcom: ipq9574: Specify probe-order for RDP467
RDP467 (AL02-C20) follows a unique WSI order that requires
PCI radios to be brought up in the order of PCI0, PCI3, PCI2
and PCI1. Up until now, probe order was being specified in the
bootargs as 0x2341 (for the above order of bring up of PCI
radios). This change is to fix the probe order in the DTS
in order to avoid specifying it from the bootargs everytime.

Change-Id: I96b69d047ac3d671ad204cd0f99bf660f7ae7ef6
Signed-off-by: Sajini R <quic_sajinir@quicinc.com>