platform: ipq: Update correct IPC handle for version commit
The software list-based ARB update should be performed instead of the
overall ARB update. Previously, during a sys upgrade, the image version
was updated in the fuse regardless of the IPC invocation due to the
previous invocation. This commit updates the IPC handle for performing the
sysupgrade, ensuring that the image version is updated based on the
IPC invocation.

Change-Id: I9b81e63e91c0b5ba7f00fcc121fce631261b0ac6
Signed-off-by: Poovendhan Selvaraj <quic_poovendh@quicinc.com>