remoteproc: qcom_q6v5_mpd: Change udelay to usleep_range
When rootpd is started after a recovery, q6_wcss_start schedules a
kthread to run handle_upd_in_rpd_crash function. In
handle_upd_in_rpd_crash function, there is a while (1) loop waiting for
rproc->state to switch to RUNNING and udelay is used, which is a
blocking wait.

In a scenario where the kthread_run schedules the kthread in the same
core where the rproc_start after recovery is happening, after
kthread_run, rproc_start notifies all subdevices of
SUBSYSTEM_AFTER_POWERUP notification. There is a msleep of 100ms in the
subdev notifier. During this msleep, the userpd kthread starts and gets
into a blocking wait leading to control never exiting the subdev
notifier and never setting rproc-state to RUNNING. This causes an RCU
stall.

Change the udelay to usleep_range to avoid blocking the execution on the
while (1) loop.

Change-Id: I87869579600b926d16e80fc8b229e7a48989b16f
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>