drivers: bus: mhi: test: Fix RDDM segment length for dump_desc alloc
Number of dump segments in the dump_desc is allocated based on the
rddm_segment length. It is assumed to be default 512K and updated based
on the DT property qti,rddm-seg-len.

However, in Linux 6.x, qti,rddm-seg-len is deprecated and RDDM segment
size is fixed to 4K. Fix the same while allocating dump_desc also.

Change-Id: I1987d0d0da2ff7f7afd3755f569112e513c0ead9
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>