drivers: clk: qcom: ipq5424: Update the freq table for sdcc1_apps clock
The divider values in sdcc1_apps freq table is wrongly updated
considering the frequency of gpll2_out_main to be 1152MHz.
But the freq of gpll2_out_main clock is 576MHz (gpll2/2) in ipq5424.
Hence revert back the divider values in the freq table to match with
the autogenerated driver.

Change-Id: I75ff24afe03e97a6152c8f3795eb19bcdd89cf6f
Signed-off-by: Devi Priya <quic_devipriy@quicinc.com>