soc: qcom: lm: Fixing the SoC HW feature list initialization
Currently, the SoC hardware feature list is initialized only if the
target is TMEL bounded. This causes the list pointers to be NULL in
case of non-TMEL bounded targets and causes kernel crash while trying
to iterate the list for exporting the FID status to softsku app.
Fixed the issue by initializing the list irrespective of TMEL bound
status.

Also, added the cleanup mechanism to free the entries in the list
as part of the driver remove call.

Change-Id: Ic9132bace01b0623f6c19247ea0e89f5705f6793
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>