platform: ipq: pass size parameter for sec.elf
The fuseipq scm with sec.elf requires size to be
passed for security improvement. The sec.dat version
of the scm doesn't require the size. Hence added the
size explicitly for elf version of the fuseipq scm.

The fuse-blow-size-required DT property is used
to determine if size has to be passed to the fuseipq
scm. But since only sec.elf needs the size parameter,
there is no other generic usecase. So removing this
property.

Change-Id: I9117b9d6dd01e800ccb440fc3a03a3391ac1073b
Signed-off-by: Kathiravan Thirumoorthy <quic_kathirav@quicinc.com>
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>