arm64: dts: qcom: ipq5424: update the probe order property in DTS file
The Q6 will be bringup based on the wifi nodes in DTS file. However, in
MLO(Multi Link Operation) enabled RDPs, we need to bringup the Q6 in
WSI(Wireless Serial Interface) order. We already have the provision to
change the probe order by using either probe_order in bootargs or
probe-order in DTS file.

Use the probe-order property in DTS file to change the bringup order
based on the WSI order. Bootargs is the higher priority then the
DTS values. If probe_order is specified in bootargs, then it overrides
the default DTS entries.

Change-Id: Id0cf7c62c86a9c2ad2fa0d29cfae4728c9d92e5d
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>