ctx-save: Add locking and wait mechanism in do_minidump
Currently WLAN driver is invoking do_minidump multiple times for each
radio. In a case where multiple calls to do_minidump is not serialized,
the device file created in the first call gets deleted and new device is
created in do_minidump. However, the first device_file is still open from
userspace. This causes a "duplicate file" warning. Also, when release
is called for the first file, the class associated with it was already
destroyed leading to a use after free case.

To avoid this, add locking mechanism to do_minidump to ensure
calls are serialized. Also add completion mechanism to wait_for_completion
after creating the file and end the wait when the file is released.

Change-Id: Ia52bfd7d928151dfad2e144814c5297c51c77684
Signed-off-by: Poovendhan Selvaraj <quic_poovendh@quicinc.com>