misc: qseecom: Updating documentation for AES type parameter
The aes_type and type_aes sysfs entries are used to determine
the destination slots to be used within CE5/TCSR regs for
distributing the key during AESv2 encrypt/decrypt scm calls.
Currently the info is updated incorrectly in documentation as
the seed value to be used. Updated the information properly in
this patch.

Change-Id: Iace81e7e6f5747348b4b485978ebcfc699ff471a
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>