misc: qseecom: Support for dynamic TZApp Log size
This change introduces a new sysfs entry 'log_size'
to determine the size for TZApp log buffer from user
before loading it. This is done to view the full
results of certain TZApp operations which usually
gets truncated due to default 4K size. The change
also falls back to legacy 4k buffer size if the
log size is not provided or if sufficient memory
cannot be allocated.

Change-Id: If33ec2206b6882fffd8db2b5c3d69b7a6bca4f1d
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>