drivers: bus: mhi: test: Change QRTR Instance ID register for QCN9625
MHI test driver programs the qrtr_instance_id base to
PCIE_PCIE_LOCAL_REG_PCIE_LOCAL_RSV0 register which the Q6 will use as
base for calculating port number to be used for bringing up the QMI
service.

The offset of this register has changed for QCN9625. Add changes to
handle this offset change for QCN9625.

Change-Id: Icb9fa7567f0cbadd069662e59cb9a58c487c4568
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>