platform: Enable rootfs authentication based on DTS fixup flag from uboot
Rootfs authentication sysfs  is currently updated based on fuse.
For boards that are deployed with FEC enable bit, rootfs auth fuse
cannot be updated anymore, thereby the feature cannot be enabled.

To fix this, it can be enabled via CONFIG in uboot, which then
patches kernel DTS with a flag. Kernel driver checks either
fuse or DTS property rootfs_auth_enable and enables the feature.

Change-Id: I15721722a02fafa0a925dff5bd32fc714dc0e0e2
Signed-off-by: Gurumoorthy Santhakumar <quic_gsanthak@quicinc.com>