qcom: dtsi: Disable ppe offload for RDP454 1G ports.
Currently, both 1G and 10G ports support PPE offload in
RDP 454. However, due to an MHT-related issue, the PPE rings
experience backpressure caused by the slowness of the
Ethernet ports. As a result, the 10G port begins to
perform similarly to a 1G port, leading to the throughtput
drops and higher EDMA TX ring size. Disabling PPE offload
for the 1G port resolves the problem.

Therefore, disabling PPE offload for the 1G port
in RDP454—which was previously missed—will help mitigate the issue.

Change-Id: I54ec7cec4e3c836736d04ec7d7f92c7409bc643b
Signed-off-by: Swati Singh <quic_swasing@quicinc.com>