drivers: qcom: ctx-save: Include crash type in minidump metadata list
This change introduces a crash type field to each minidump
TLV and metadata structure.

The crash type will be used to determine which segments to dump,
enabling selective dumping based on the cause of the system crash
from dmesg.

Additionally, this change introduces support for up to 1024 TLV entries
in the metadata.

Change-Id: Ifb50a37a3988588981056a097cf7eb024bf72e2e
Signed-off-by: kondappa <kondappa@qti.qualcomm.com>