arm64: dts: qcom: Add mhi-region to memory-region in RDP-479,RDP-481.
Currently in RDP-479 & RDP-481 the memory-region node of WKK have only
host-ddr-mem entry, mlo-global-mem entry. But in case of prop driver
if at all there exist an entry in 1st index of memory-region node that
memory region will be used to load amss.bin because of which amss.bin
is being loaded into mlo-global-mem which is causing a bootup crash.

Eariler as there was only 1 entry at index-0 mhi was using common memory
pool to load amss.bin now with entry in index-1 as mlo-global-mem mhi
will try to consume that. As the other entries are needed for ATH driver
adding mhi-region to index-1 and moving mlo-global-mem to index-2.

Fixes: I9af9cb6ac (arm64: dts: qcom: Represent WSI interconnection with TX/RX ports)

Change-Id: Ieb00f3e446b65a1b71064ab2eeb42493bdfda8d4
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>