arm64: dts: qcom: ipq5332: update the phandle for RDP479 and RDP481
Currently, the early_cal property is added in base DTSi file for each
wifi nodes and its pointing to the PCIx node for getting the Early
calibration flag. In RDP sepcific DTS file, the wifi node is pointing
to different PCIx node and thus Early cal is not enabled.

To fix this issue, override the early_cal property in RDP sepcific DTS
file to point to the correct PCIx node for getting the Early cal flag.

Change-Id: I92cd94e5e050ca9332e4fd31a90afd57ee7599de
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>