Added a check to disable checksum calculation for list head
When skb debug objects are enabled in KASAN build and a flush
command is exucuted on skb recycle list, 'stack out of bound'
message was reported by KASAN along with call trace. This is
because the function that calculates the byte-wise sum of skb
got a pointer of type skbuff_head instead of type skb after the
last skb was unliked from the list.

This patch does avoids sending skbuff_head pointer.

Change-Id: I18584964f2420c919d07f01a14bd7a1e62ddd710
Signed-off-by: Amruth S <amrus@qti.qualcomm.com>