arm64: dts: ipq9574: Disable Early Cal for SDX RDPs
From u-boot, we cannot control the PCIe slots to which PERST signal is
issued. So PERST gets issued for SDX also, which causes SDX to power on
and go to bad state since MHI stack in kernel is not initialized yet.

Disable Early Cal support for SDX RDPs RDP461 and RDO461-C21.

Change-Id: I02ef3828a228597af93df8dc47871b944f76f0f3
Signed-off-by: Vignesh Viswanathan <viswanat@qti.qualcomm.com>