arm64: dts: qcom: ipq: Modify the property for early cal enablement
Currently, qcom,early_cal_enabled is set to "okay" by default.
If the bootloader is not able to start the early cal, it will patch
the device tree to set it to "disabled".

However, when the bootloader does not support early cal, it does not
patch this property and does not start the calibration, but the pcie
driver thinks calibration is enabled, times out and panics.

Set qcom,early_cal_enabled to "supported" by default. Now bootloader
will check if this property is set to "supported", try to start the
calibration and patch it to "okay" only if it starts the calibration
successfully.

To disable early cal, this property can be set to "disabled".

Change-Id: If53307bfef0620710a836101c8ac999617a0b312
Signed-off-by: Vignesh Viswanathan <viswanat@qti.qualcomm.com>