arm64: dts: qcom: Represent WSI interconnection with TX/RX ports
All WiFi devices that are attached to an RDP are inter connected
among themselves via  WSI connection. This WSI connection is
essential to exchange control information among these devices.
Currently it is being reperesnted in below format in wsi Node.

    <chip_idx num_adj_chips adj_chip_idx[0] adj_chip_idx[1]>

Instead of this defining TX/RX ports in each wifi node to
represent the inter connextion.

For Example The WSI connection in RDP433 is represented below:

                     +-------+        +-------+        +-------+
                     | pcie1 |        | pcie2 |        | pcie3 |
                     |       |        |       |        |       |
             +------>|  wsi  |------->|  wsi  |------->|  wsi  |-----+
             |       | grp 0 |        | grp 0 |        |grp 0  |     |
             |       +-------+        +-------+        +-------+     |
             +-------------------------------------------------------+

Based on the above, the wifi  node properties for QCN9274 at
pcie2 are (considering QCN9274 at pcie2 is WSI controller):

- qcom,wsi-controller
- ports:
        tx-port (port@0):endpoint at pcie3 RX port.
        rx-port (port@1):endpoint at pcie1 TX port.

Also for fixed memory mode added a new property memory-region-names
which will be one to one mapping to the property memory-region and
these names will be used in ath12k driver to fetch a particular
memory location instead of a fixed index.

To not break the mainline adding q6_region in both index0 and index1
with names q6-region,q6-region1. This needs to be cleaned up.

Change-Id: I6ad91e26579d051c2b516c63961cc90bab8835b4
Signed-off-by: Narendhar Reddy <nare@qti.qualcomm.com>