drivers: pci: ipq52xx: Add SoC register configuration
This patch adds support for configuring and accessing
SoC registers on IPQ52xx platforms through the FPGA
connected via PCIe on IPQ95xx devices.

The configuration details for these registers are
specified using a Device Tree property called
`qcom,soc-reg-config`. This property defines a list
of triplets in the following format:

    <start size offset>

Where:
    - start:  Base physical address of the SoC region
    - size:   Size of the region in bytes
    - offset: Offset within the PCI BAR where the
              region is mapped

When this property is present in the DTS, the driver
uses these mappings to calculate BAR offsets for SoC
register accesses dynamically.

The DTS must also include the boolean property
`enable-soc-reg-config` to explicitly enable this
feature.

Change-Id: Ifbc2b56c4e41206a318b8857ee55d988854c3993
Signed-off-by: Timple Raj M <timple@qti.qualcomm.com>